requirements/requirements.txt
asgiref
certifi
charset-normalizer
Django
djangorestframework
idna
pytz
requests
sqlparse
urllib3
django-environ
mysqlclient
django-mysql